Все классы переписаны почти с нуля. Куча говнокода, куча хардкода -- одним словом это дебаг-версия для теста нового подхода к модели. Я отказался от чисел с запятой, теперь только целые числа. Устранены зависания при беге, бег и прыжки стали реалистичнее.

+++ WindowsFormsApp1/WindowsFormsApp1/Demon.cs
@@ -9,7 +9,7 @@ namespace WindowsFormsApp1
 {
     class Demon : Enemy
     {
-        public Demon(int HP, Vector location, double width, double height, Bitmap sprite) : base(HP, location, width, height, sprite)
+        public Demon(int HP, Vector location, int width, int height, Bitmap sprite) : base(HP, location, width, height, sprite)
         {
         }
     }
