+++ WindowsFormsApp1/WindowsFormsApp1/Player.cs
@@ -4,7 +4,7 @@ namespace WindowsFormsApp1
 {
     public class Player : Entity
     {
-        public Player(int HP, Vector location, double width, double height, Bitmap sprite) : base(HP, location, width, height, sprite)
+        public Player(int HP, Vector location, int width, int height, Bitmap sprite) : base(HP, location, width, height, sprite)
         {
 
         }
